pyproject.toml

[project]
authors = [
  { name = "Angel L. P." },
  { name = "chrisAta" },
  { name = "MGS-sails" },
  { name = "mberacochea" },
  { name = "SandyRogers" },
  { name = "Sebastian Ayala Ruano" },
  { name = "Alberto Santos" }
]
description = "python driver for mgnipy api"
name = "mgnipy"
# See the section below: [tools.setuptools.dynamic]
dynamic = [
  "version",
]
readme = "README.md"
requires-python = ">=3.10" 
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
license = "GPL-3.0-or-later" # https://choosealicense.com/licenses/gpl-3.0/
dependencies = [
    "anndata>=0.10.9",
    "attrs>=25.4.0",
    "httpx>=0.28.1",
    "numpy>=2.0.2",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scipy>=1.13.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "black[jupyter]>=25.11.0",
    "isort>=6.1.0",
    "openapi-python-client>=0.17.3",
    "pytest>=8.4.2",
    "ruff>=0.14.13",
]
docs = [
    "ipykernel>=6.31.0",
    "ipywidgets>=8.1.8",
    "jupytext>=1.19.0",
    "myst-nb>=1.3.0",
    "sphinx>=7.4.7",
    "sphinx-book-theme>=1.1.4",
    "sphinx-copybutton>=0.5.2",
    "sphinx-new-tab-link>=0.8.0",
]

[tool.isort]
profile = "black"
multi_line_output = 3
force_grid_wrap = 2
include_trailing_comma = true
use_parentheses = true

[tool.ruff.lint]
# https://docs.astral.sh/ruff/tutorial/#rule-selection
extend-select = ["E", "W", "F", "B"]
ignore = ["E501"]  # ignore line length for now

[tool.pytest.ini_options]
addopts = "--doctest-modules"

[tool.setuptools.packages.find]
include = ["mgnipy*"]

[tool.setuptools_scm]

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64", "setuptools_scm>=8"]

[project.scripts]
mgnipy-hello = "mgnipy.cli:main"

[project.urls]
"Bug Tracker" = "https://github.com/EBI-Metagenomics/mgnipy/issues"
"Homepage" = "https://github.com/EBI-Metagenomics/mgnipy"
